feat: impl all hardware controller, service, repo first iteration

+++ internals/models/domain/hardware.go
@@ -1,7 +1,7 @@
 package domain
 
 type Hardware struct {
-	IdHardware  int    `json:"id_hardware"`
+	IdHardware  int64  `json:"id_hardware"`
 	Name        string `json:"name"`
 	Type        string `json:"type"`
 	Description string `json:"description"`
